+++ packages/frontend/components/collection/layouts/magazine/MagazineItem.tsx
@@ -11,6 +11,7 @@ import {
   BoxProps,
 } from '@chakra-ui/react';
 import NextLink from 'next/link';
+import { format, fromUnixTime } from 'date-fns';
 import { CollectionItemProps } from '../../types';
 
 export type MagazineItemProps = CollectionItemProps & BoxProps;
@@ -25,6 +26,7 @@ export const MagazineItem = forwardRef((props: MagazineItemProps, ref) => {
     collection,
     readingTime,
     dateRead,
+    datePublished,
   } = item;
   const readingTimeRounded = Math.ceil(readingTime);
 
@@ -64,7 +66,8 @@ export const MagazineItem = forwardRef((props: MagazineItemProps, ref) => {
             {summary}
           </Text>
           <Text color="gray.500">
-            by {collection.title}
+            by {collection.title} •{' '}
+            {format(fromUnixTime(datePublished), 'dd/MM/yyyy')}
             {readingTimeRounded > 0 && ` • about ${readingTimeRounded} min`}
           </Text>
         </VStack>